Fix .mpack not working in VS Mac
Ensure the vsmac-addin-compat.runtimeconfig.json file is included
in the .mpack. This allows the vsmac-addin-compat.dll to be run with
dotnet. Fixes this error reported in the IDE log:

A fatal error was encountered. The library 'libhostpolicy.dylib' required to execute the application was not found in '~/Library/Application Support/VisualStudio/17.0/LocalInstall/Addins/Microsoft.VisualStudioMac.ExtensionCompatibilityChecker.0.1/'.
Failed to run as a self-contained app.
  - The application was run as a self-contained app because '~/Library/Application Support/VisualStudio/17.0/LocalInstall/Addins/Microsoft.VisualStudioMac.ExtensionCompatibilityChecker.0.1/vsmac-addin-compat.runtimeconfig.json' was not found.
  - If this should be a framework-dependent app, add the '~/Library/Application Support/VisualStudio/17.0/LocalInstall/Addins/Microsoft.VisualStudioMac.ExtensionCompatibilityChecker.0.1/vsmac-addin-compat.runtimeconfig.json' file and specify the appropriate framework.