Ensure all temporary report files are removed
The binary compat checker creates more than just the report file.
It generates an assembly list file and two ivt list files. To
ensure these are not left behind a temporary directory is created,
instead of a temporary file, and then the directory is removed
after the check is run.