Ignore compat problems from VS Mac itself
Generate a baseline for VS Mac first, before checking addin compat. This
VS Mac baseline is used to ignore compat problems that exist in VS Mac
itself and are not specific to the addin being checked.

Removed the --addin-baseline-file option. No longer going to support
this.

Added a --generate-vsmac-baseline option. This is used to support
generating the vsmac baseline so it can be re-used.