Support generating a compat diff file and also ignoring known diffs
The vsmac-addin-compat command line tool now supports two new
arguments:

-compat-diff-output-file
-compat-diff-ignore-file

compat-diff-output-file is used to specify a filename where the addin
compat diff line information will be saved to. These only include the
added lines when using VS Mac as the baseline.

The compat-diff-ignore-file is used to specify the filename containing
diff lines that can be ignored when running the compat check.

The idea behind these two arguments is to allow known compat errors
to be ignored.