Allow extension baselines to be saved and reset
The extension compatibility addin now has two buttons in its
Extension Compatibility Console pad window. 'Save baselines' will
save the current diffs reported in the pad window so they are ignored
the next time the IDE is opened. The 'Reset baselines' will remove
these saved diffs so they are not used on re-opening the IDE. This
allows known addin compat problems to be ignored but allows new
problems to be reported, such as if the IDE's public API has changed.